Updated saved block content and JS to be unobtrusive
While developing the Eventbrite block we hit upon a pattern that would
allow the button HTML to be saved as part of the post content. The JS
code could then be hooked based on the element ID rather than using a
click handler. It also meant that the HTML would function as a standard
link if the JS wasn't loaded - the block was unavailable or it was an
AMP request, email etc.